Responded to the first round of comments
