+++ src/main/java/org/broadinstitute/hellbender/utils/read/SAMRecordToGATKReadAdapter.java
@@ -719,16 +719,10 @@ public class SAMRecordToGATKReadAdapter implements GATKRead, Serializable {
     }
 
     public Object getTransientAttribute(Object key) {
-        if (samRecord==null) {
-            throw new GATKException("Underlying sam record cannot be null when setting transient attributes");
-        }
         return samRecord.getTransientAttribute(key);
     }
 
     public void setTransientAttribute(Object key, Object value) {
-        if (samRecord==null) {
-            throw new GATKException("Underlying sam record cannot be null when setting transient attributes");
-        }
         samRecord.setTransientAttribute(key, value);
     }
 }
