+++ src/main/java/org/broadinstitute/hellbender/utils/read/markduplicates/sparkrecords/PairedEnds.java
@@ -1,0 +1,18 @@
+package org.broadinstitute.hellbender.utils.read.markduplicates.sparkrecords;
+
+/**
+ * Struct-like class to store information about the paired reads for mark duplicates.
+ */
+public abstract class PairedEnds extends MarkDuplicatesSparkRecords {
+
+    PairedEnds(int partitionIndex, String name) {
+        super(partitionIndex, name);
+    }
+
+    public abstract int getFirstStartPosition();
+    public abstract int getUnclippedStartPosition();
+    public abstract int getFirstRefIndex();
+    public abstract boolean isR1R();
+    public abstract int getScore();
+
+}
